remove cycle in skeletontracing_reducer_helpers
